separated class
